Rewrite the readme
I have a feeling this book just isn't for me. Haskell might be a great language but there's a reason it's niche and if it's written by mathematicians and overly academic, it's, again, not for me.

I want a lighter language that's good for building basic programs and prototypes, not a heavy backend language however good it is.